Correct definition of CLIPFORMAT.
--HG--
extra : convert_revision : svn%3Aa2f44796-8cc0-49ac-b43f-6a96d556d52d/trunk%40125

+++ comtypes/persist.py
@@ -11,6 +11,7 @@ The 'DictPropertyBag' class is a class implementing the IPropertyBag
 interface, useful in client code.
 """
 from ctypes import *
+from ctypes.wintypes import WORD
 from comtypes import GUID, IUnknown, COMMETHOD, HRESULT, dispid
 from comtypes.automation import VARIANT, tagEXCEPINFO
 
@@ -67,7 +68,7 @@ class IPersistPropertyBag(IPersist):
         ]
 
 
-CLIPFORMAT = c_ulong # Hm, is that correct?
+CLIPFORMAT = WORD
 
 PROPBAG2_TYPE_UNDEFINED = 0
 PROPBAG2_TYPE_DATA = 1
